Struct limn_core::event::EventArgs

pub struct EventArgs<'a> {
    pub widget: Widget,
    pub ui: &'a mut Ui,
    pub handled: &'a mut bool,
}

Context passed to a EventHandler, provides access to the widget that holds it, the Ui, and a flag to notify the dispatcher that the event has been handled (in the case the event is bubbling up)

Fields
widget: Widget ui: &'a mut Ui handled: &'a mut bool