Module limn_core::geometry

Geometric type definitions, such as Point, Rect and Size

Structs

DevicePixel
Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

Traits

RectExt
Extension trait for rectangles. Helper methods for rectangle sides depend on the assumption that Rect size never contains negative values.

SizeExt
Extension trait for sizes.

Type Definitions

DensityIndependentPixel
This is the unit used for all geometric types that represent a position on screen, independent of the device's hidpi factor, or "Device Pixel Ratio". Sometimes known as 'points', logical pixels or device independent pixels,

Point

Rect

Size

Vector