Struct limn_core::input::keyboard::FocusHandler

pub struct FocusHandler { /* fields omitted */ }

Note on focus: The tab key iterates through the widgets that have registered as focusable. Currently the order of this iteration is just based on the order the widgets are registered as focusable. Later on maybe it should be based on the relative positioning of widgets (could get ugly updating the treemap as widgets change position), or some user defined ordering.

Methods

impl FocusHandler

pub fn new() -> Self