Struct limn_core::prelude::Constraint

pub struct Constraint(_);

A constraint, consisting of an equation governed by an expression and a relational operator, and an associated strength.

Methods

impl Constraint

pub fn new(e: Expression, op: RelationalOperator, strength: f64) -> Constraint

Construct a new constraint from an expression, a relational operator and a strength. This corresponds to the equation e op 0.0, e.g. x + y >= 0.0. For equations with a non-zero right hand side, subtract it from the equation to give a zero right hand side.

pub fn expr(&self) -> &Expression

The expression of the left hand side of the constraint equation.

pub fn op(&self) -> RelationalOperator

The relational operator governing the constraint.

pub fn strength(&self) -> f64

The strength of the constraint that the solver will use.