Struct limn_core::ui::Ui

pub struct Ui {
    pub window: Rc<RefCell<Window>>,
    // some fields omitted
}

The core of a limn application, holds the root of the widget tree and other application global state. Ui is accessible to every event handler, so features helper methods that can be accessed at any time.

Fields
window: Rc<RefCell<Window>>
Methods

impl Ui

pub fn get_widget(&self, widget_id: WidgetId) -> Option<Widget>

pub fn get_root(&self) -> Widget

pub fn event<T: 'static>(&self, data: T)

pub fn close(&mut self)

pub fn check_layout_changes(&mut self)

pub fn redraw(&mut self)

pub fn needs_redraw(&self) -> bool

Important traits for WidgetsBfs
impl Iterator for WidgetsBfs type Item = Widget;

pub fn widgets_bfs(&self) -> WidgetsBfs

Important traits for WidgetsUnderCursor
impl Iterator for WidgetsUnderCursor type Item = Widget;

pub fn widgets_under_cursor(&mut self, point: Point) -> WidgetsUnderCursor

pub fn widget_under_cursor(&mut self, point: Point) -> Option<Widget>

Find the first widget under the cursor, ie. the last to be drawn that is under the cursor

pub fn set_debug_draw_bounds(&mut self, debug_draw_bounds: bool)

pub fn debug_widget_positions(&self)

pub fn print_widgets(&self)