Module limn_core::widget

Module for Widget and callback handlers Types used to create and manage widgets.

Limn UIs exist as a tree of Widgets, each of which consists of a bounding rectangle, a list of references to it's children, a list of EventHandlers that receive and send events, and optionally a draw state struct that implements Draw.

The root widget is just an ordinary widget that happens to be stored by the Ui so it can be drawn, and that has the size of the window as its bounding rectangle.

Creating a user interface consists of constructing a widget tree, then passing the Widget root of that tree to a limn App, which will attach it to the Ui root widget for you and begin the event loop.

Modules

draw

filter

property

Structs

DrawStateGuard

LayoutGuard

LayoutGuardMut

PropsGuard

StateUpdated

StyleUpdated

Widget

WidgetWeak