Trait limn_core::widget::filter::Filter

pub trait Filter: Any {
    fn push(&self, renderer: &mut RenderBuilder);

    fn pop(&self, renderer: &mut RenderBuilder);
}

Required Methods

fn push(&self, renderer: &mut RenderBuilder)

fn pop(&self, renderer: &mut RenderBuilder)

Methods

impl Filter

pub fn is<T: Filter>(&self) -> bool

Returns true if the boxed type is the same as T

pub fn downcast_ref<T: Filter>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn downcast_ref_unchecked<T: Filter>(&self) -> &T

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

pub fn downcast_mut<T: Filter>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn downcast_mut_unchecked<T: Filter>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl Filter

pub fn downcast<T: Filter>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Returns the boxed value if it is of type T, or Err(Self) if it isn't.

Important traits for Box<I>
impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub unsafe fn downcast_unchecked<T: Filter>(self: Box<Self>) -> Box<T>

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.