Struct limn_core::window::Window

pub struct Window {
    pub window: GlWindow,
}

A simple wrapper around a glutin::GlWindow.

Fields
window: GlWindow
Methods

impl Window

pub fn new(window_builder: WindowBuilder, events_loop: &EventsLoop) -> Self

pub fn gl(&self) -> Rc<Gl>

pub fn swap_buffers(&self)

pub fn hidpi_factor(&self) -> f32

pub fn resize(&mut self, width: u32, height: u32)

pub fn size_px(&self) -> DeviceUintSize

Get the size of the client area of the window in actual pixels. This is the size of the framebuffer

pub fn size_dp(&self) -> Size

Get the size of the client area of the window in density independent pixels.

pub fn show(&self)